requirements.txt
# Generic Pipeline Dependencies

# Data processing
pandas>=2.0.0

# LLM API
google-generativeai>=0.3.0

# Schema validation
pydantic>=2.0.0

# Progress tracking
tqdm>=4.65.0

# Environment management
python-dotenv>=1.0.0

